feat: get user

+++ src/main/java/br/com/raroacademy/demo/domain/DTO/UserResponseDTO.java
@@ -3,5 +3,5 @@ package br.com.raroacademy.demo.domain.DTO;
 import lombok.Builder;
 
 @Builder
-public record UserResponseDTO(Integer id, String name, String email, Boolean emailConfirmed) {
+public record UserResponseDTO(Long id, String name, String email, Boolean emailConfirmed) {
 }
